feat: implementing heap_min_priority

+++ heap_testr/heap.h
@@ -15,7 +15,7 @@ void* heap_push(Heap *heap, void *data, double priority);
 bool heap_empty(Heap *heap);
 void* heap_max(Heap *heap);
 double heap_max_priority(Heap *heap);
-int heap_min_priority(Heap* heap);
+double heap_min_priority(Heap* heap);
 void* heap_pop(Heap *heap);
 void heap_destroy(Heap *heap);
